pyproject.toml
[project]
name = "arxiv-rag"
version = "0.1.0"
description = "Data analysis bot"
requires-python = ">=3.11"
# Added dependencies here
dependencies = [
    "python-multipart==0.0.20",
    "pydantic==2.11.7",
    "gunicorn==23.0.0",
    "uvicorn==0.34.2",
    "python-dotenv==1.1.0",
    "tenacity==9.1.2",
    "requests==2.32.5",
    "pytz==2025.2",
    "httpx==0.28.1",
    "langgraph==0.6.6",
    "langchain==0.3.27",
    "langchain-openai==0.3.32",
    "ipython==9.2.0",
    "asyncpg==0.30.0",    
    "aiohttp==3.11.18",           
    "aiofiles==24.1.0",          
    "langchain-community==0.3.29", 
    "langchain-experimental==0.3.4",
    "langchain-text-splitters",
    "flask",
    "flasgger",
    "deepeval"
 
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",    
    "pytest-cov==6.1.1",
    "pytest-asyncio==1.1.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"